requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
flask==3.0.0
requests==2.31.0
firebase-admin==6.4.0
aiohttp==3.9.1
